+++ wasmer-it/src/interpreter/instructions/byte_arrays.rs
@@ -33,11 +33,11 @@ executable_instruction!(
                     )
                 })?;
 
-            let pointer: usize = to_native::<i32>(inputs.remove(0), instruction.clone())?
+            let pointer: u32 = to_native::<i32>(inputs.remove(0), instruction.clone())?
                 .try_into()
                 .map_err(|e| (e, "pointer").into())
                 .map_err(|k| InstructionError::from_error_kind(instruction.clone(), k))?;
-            let length: usize = to_native::<i32>(inputs.remove(0), instruction.clone())?
+            let length: u32 = to_native::<i32>(inputs.remove(0), instruction.clone())?
                 .try_into()
                 .map_err(|e| (e, "length").into())
                 .map_err(|k| InstructionError::from_error_kind(instruction.clone(), k))?;
@@ -74,12 +74,12 @@ executable_instruction!(
                 )
             })?;
 
-            let array_pointer: usize = to_native::<i32>(inputs.remove(0), instruction.clone())?
+            let array_pointer: u32 = to_native::<i32>(inputs.remove(0), instruction.clone())?
                 .try_into()
                 .map_err(|e| (e, "pointer").into())
                 .map_err(|k| InstructionError::from_error_kind(instruction.clone(), k))?;
             let array: Vec<u8> = to_native(inputs.remove(0), instruction.clone())?;
-            let length: i32 = array.len().try_into().map_err(|_| {
+            let length: u32 = array.len().try_into().map_err(|_| {
                 InstructionError::from_error_kind(
                     instruction.clone(),
                     InstructionErrorKind::NegativeValue { subject: "array_length" },
@@ -99,14 +99,14 @@ executable_instruction!(
                 .view();
 
             let writer = memory_view
-                .sequential_writer(array_pointer, array.len())
+                .sequential_writer(array_pointer, array.len() as u32)
                 .map_err(|e| InstructionError::from_memory_access(instruction.clone(), e))?;
 
             writer.write_bytes(&array);
 
             log::debug!("string.lower_memory: pushing {}, {} on the stack", array_pointer, length);
             runtime.stack.push(IValue::I32(array_pointer as i32));
-            runtime.stack.push(IValue::I32(length));
+            runtime.stack.push(IValue::I32(length as i32));
 
             Ok(())
         }
